refactor(memory): ChipArena (chip en el tipo) + MemorySystem tipada; Scene/Bitmap sin puente runtime

+++ tests/host/field/045_surface_polygon/src/main.cpp
@@ -27,8 +27,8 @@ struct MockPlayfield : Playfield {
 		m_frontbuffer = bank.reserve<eng::PlaneTag>(m_total_bytes, 2u).address();
 		m_initialized = true;
 	}
-	u32 planeline_for(s32 wy) const override { return static_cast<u32>(wy) * m_planes; }
-	u32 byte_for(s32 wx) const override { return (static_cast<u32>(wx) / 8u) & ~1u; }
+	u32 planeline_for(s32 wy) const override { return wy * m_planes; }
+	u32 byte_for(s32 wx) const override { return wx / 8 & ~1; }
 	bool in_bounds(s32 wx, s32 wy) const override {
 		return wx >= 0 && wy >= 0 && static_cast<u32>(wx) < m_width && static_cast<u32>(wy) < m_height;
 	}
